Updated tests based on comments provided.
