+++ tests/components/elevenlabs/conftest.py
@@ -14,8 +14,11 @@ from homeassistant.components.elevenlabs.const import (
     CONF_STT_AUTO_LANGUAGE,
     CONF_STT_MODEL,
     CONF_VOICE,
+    DEFAULT_SIMILARITY,
+    DOMAIN,
 )
 from homeassistant.const import CONF_API_KEY
+from homeassistant.core import HomeAssistant
 
 from .const import MOCK_MODELS, MOCK_VOICES
 
@@ -142,6 +145,55 @@ def mock_async_client_connect_error() -> Generator[AsyncMock]:
         yield mock_async_client
 
 
+async def mock_config_entry_setup(
+    hass: HomeAssistant, config_data: dict[str, Any], config_options: dict[str, Any]
+) -> None:
+    """Mock config entry setup."""
+    default_config_data = {
+        CONF_API_KEY: "api_key",
+    }
+    default_config_options = {
+        CONF_VOICE: "voice1",
+        CONF_MODEL: "model1",
+        CONF_STT_MODEL: "stt_model1",
+        CONF_STT_AUTO_LANGUAGE: False,
+    }
+    config_entry = MockConfigEntry(
+        domain=DOMAIN,
+        data=default_config_data | config_data,
+        options=default_config_options | config_options,
+    )
+    config_entry.add_to_hass(hass)
+    client_mock = AsyncMock()
+    client_mock.voices.get_all.return_value = GetVoicesResponse(voices=MOCK_VOICES)
+    client_mock.models.list.return_value = MOCK_MODELS
+    stt_mock = AsyncMock()
+    stt_mock.convert.return_value = AsyncMock(
+        text="hello world", language_code="en", language_probability=0.95
+    )
+    client_mock.speech_to_text = stt_mock
+    with patch(
+        "homeassistant.components.elevenlabs.AsyncElevenLabs", return_value=client_mock
+    ):
+        assert await hass.config_entries.async_setup(config_entry.entry_id)
+
+
+@pytest.fixture(name="setup")
+async def setup_fixture(
+    hass: HomeAssistant,
+    config_data: dict[str, Any],
+    config_options: dict[str, Any],
+    mock_async_client: AsyncMock,
+) -> AsyncMock:
+    """Set up the test environment."""
+
+    await mock_config_entry_setup(hass, config_data, config_options)
+
+    await hass.async_block_till_done()
+
+    return mock_async_client
+
+
 @pytest.fixture
 def mock_entry() -> MockConfigEntry:
     """Mock a config entry."""
@@ -176,3 +228,9 @@ def config_data_fixture() -> dict[str, Any]:
 def config_options_fixture() -> dict[str, Any]:
     """Return config options."""
     return {}
+
+
+@pytest.fixture
+def mock_similarity():
+    """Mock similarity."""
+    return DEFAULT_SIMILARITY / 2
